LPS-115840 Move NoSuchClassTypeException to com.liferay.info.exception
