+++ modules/apps/info/info-api/bnd.bnd
@@ -8,6 +8,7 @@ Export-Package:\
 	com.liferay.info.display.field,\
 	com.liferay.info.display.request.attributes.contributor,\
 	com.liferay.info.display.url.provider,\
+	com.liferay.info.exception,\
 	com.liferay.info.field,\
 	com.liferay.info.field.type,\
 	com.liferay.info.form,\
