+++ modules/apps/info/info-api/src/main/resources/com/liferay/info/exception/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
